Split up PPOProjectCardModel into multiple files for templates and parsing
