+++ Kickstarter-iOS/Features/PledgedProjectsOverview/CardView/PPOProjectCardModel.swift
@@ -205,292 +205,3 @@ extension GraphAPI.MoneyFragment: Hashable {
     hasher.combine(self.symbol)
   }
 }
-
-extension PPOProjectCardModel {
-  #if targetEnvironment(simulator)
-    public static let previewTemplates: [PPOProjectCardModel] = [
-      confirmAddressTemplate,
-      addressLockTemplate,
-      fixPaymentTemplate,
-      authenticateCardTemplate,
-      completeSurveyTemplate
-    ]
-  #endif
-
-  internal static let confirmAddressTemplate = PPOProjectCardModel(
-    isUnread: true,
-    alerts: [
-      .init(type: .warning, icon: .time, message: "Address locks in 8 hours")
-    ],
-    image: .network(URL(string: "https:///")!),
-    projectName: "Sugardew Island - Your cozy farm shop let’s pretend this is a way way way longer title",
-    projectId: 12_345,
-    pledge: "$50.00",
-    creatorName: "rokaplay truncate if longer than",
-    address: """
-      Firsty Lasty
-      123 First Street, Apt #5678
-      Los Angeles, CA 90025-1234
-      United States
-    """,
-    actions: (.confirmAddress(
-      address: """
-        123 First Street, Apt #5678
-        Los Angeles, CA 90025-1234
-        United States
-      """,
-      addressId: "fake-address-id"
-    ), .editAddress),
-    tierType: .confirmAddress,
-    backingDetailsUrl: "fakeBackingDetailsUrl",
-    backingId: 47,
-    backingGraphId: "backing-fake-id",
-    projectAnalytics: Self.projectAnalyticsFragmentTemplate
-  )
-
-  internal static let addressLockTemplate = PPOProjectCardModel(
-    isUnread: true,
-    alerts: [
-      .init(type: .warning, icon: .alert, message: "Survey available"),
-      .init(type: .warning, icon: .time, message: "Address locks in 48 hours")
-    ],
-    image: .network(URL(string: "https:///")!),
-    projectName: "Sugardew Island - Your cozy farm shop let’s pretend this is a way way way longer title",
-    projectId: 12_345,
-    pledge: "$50.00",
-    creatorName: "rokaplay truncate if longer than",
-    address: nil,
-    actions: (.completeSurvey, nil),
-    tierType: .openSurvey,
-    backingDetailsUrl: "fakeBackingDetailsUrl",
-    backingId: 47,
-    backingGraphId: "backing-fake-id",
-    projectAnalytics: Self.projectAnalyticsFragmentTemplate
-  )
-
-  internal static let fixPaymentTemplate = PPOProjectCardModel(
-    isUnread: true,
-    alerts: [
-      .init(type: .alert, icon: .alert, message: "Payment failed"),
-      .init(
-        type: .alert,
-        icon: .time,
-        message: "Pledge will be dropped in 6 days"
-      )
-    ],
-    image: .network(URL(string: "https:///")!),
-    projectName: "Sugardew Island - Your cozy farm shop let’s pretend this is a way way way longer title",
-    projectId: 12_345,
-    pledge: "$50.00",
-    creatorName: "rokaplay truncate if longer than",
-    address: nil,
-    actions: (.fixPayment, nil),
-    tierType: .fixPayment,
-    backingDetailsUrl: "fakeBackingDetailsUrl",
-    backingId: 47,
-    backingGraphId: "backing-fake-id",
-    projectAnalytics: Self.projectAnalyticsFragmentTemplate
-  )
-
-  internal static let authenticateCardTemplate = PPOProjectCardModel(
-    isUnread: true,
-    alerts: [
-      .init(type: .alert, icon: .alert, message: "Card needs authentication"),
-      .init(
-        type: .alert,
-        icon: .time,
-        message: "Pledge will be dropped in 6 days"
-      )
-    ],
-    image: .network(URL(string: "https:///")!),
-    projectName: "Sugardew Island - Your cozy farm shop let’s pretend this is a way way way longer title",
-    projectId: 12_345,
-    pledge: "$50.00",
-    creatorName: "rokaplay truncate if longer than",
-    address: nil,
-    actions: (.authenticateCard(clientSecret: ""), nil),
-    tierType: .authenticateCard,
-    backingDetailsUrl: "fakeBackingDetailsUrl",
-    backingId: 47,
-    backingGraphId: "backing-fake-id",
-    projectAnalytics: Self.projectAnalyticsFragmentTemplate
-  )
-
-  internal static let completeSurveyTemplate = PPOProjectCardModel(
-    isUnread: true,
-    alerts: [
-      .init(type: .warning, icon: .alert, message: "Survey available")
-    ],
-    image: .network(URL(string: "https:///")!),
-    projectName: "Sugardew Island - Your cozy farm shop let’s pretend this is a way way way longer title",
-    projectId: 12_345,
-    pledge: "$50.00",
-    creatorName: "rokaplay truncate if longer than",
-    address: nil,
-    actions: (.completeSurvey, nil),
-    tierType: .openSurvey,
-    backingDetailsUrl: "fakeBackingDetailsUrl",
-    backingId: 47,
-    backingGraphId: "backing-fake-id",
-    projectAnalytics: Self.projectAnalyticsFragmentTemplate
-  )
-
-  internal static let projectAnalyticsFragmentTemplate = GraphAPI.ProjectAnalyticsFragment(
-    addOns: nil,
-    backersCount: 42,
-    backing: nil,
-    category: nil,
-    commentsCount: 42,
-    country: .init(code: .us),
-    creator: nil,
-    currency: .usd,
-    deadlineAt: nil,
-    launchedAt: nil,
-    pid: 42,
-    name: "Test",
-    isInPostCampaignPledgingPhase: true,
-    isWatched: true,
-    percentFunded: 100,
-    isPrelaunchActivated: false,
-    projectTags: [],
-    postCampaignPledgingEnabled: false,
-    rewards: nil,
-    state: .successful,
-    video: nil,
-    pledged: .init(amount: nil),
-    fxRate: 4,
-    usdExchangeRate: nil,
-    posts: nil,
-    goal: nil
-  )
-}
-
-private enum PPOProjectCardModelConstants {
-  static let paymentFailed = "Tier1PaymentFailed"
-  static let confirmAddress = "Tier1AddressLockingSoon"
-  static let completeSurvey = "Tier1OpenSurvey"
-  static let authenticationRequired = "Tier1PaymentAuthenticationRequired"
-}
-
-extension PPOProjectCardModel {
-  init?(node: GraphAPI.FetchPledgedProjectsQuery.Data.PledgeProjectsOverview.Pledge.Edge.Node) {
-    let card = node.fragments.ppoCardFragment
-    let backing = card.backing?.fragments.ppoBackingFragment
-    let ppoProject = backing?.project?.fragments.ppoProjectFragment
-
-    let image = ppoProject?.image?.url
-      .flatMap { URL(string: $0) }
-      .map { Kingfisher.Source.network($0) }
-
-    let projectName = ppoProject?.name
-    let projectId = ppoProject?.pid
-    let pledgeFragment = backing?.amount.fragments.moneyFragment
-    let formattedPledge = pledgeFragment.flatMap { Format.currency($0) }
-    let creatorName = ppoProject?.creator?.name
-
-    let addressId: String? = backing?.deliveryAddress?.id
-    let addressWithoutName: String? = backing?.deliveryAddress.flatMap { deliveryAddress in
-      let cityRegionFields: [String?] = [
-        deliveryAddress.city,
-        deliveryAddress.region.flatMap { ", \($0)" },
-        deliveryAddress.postalCode.flatMap { " \($0)" }
-      ]
-      let fields: [String?] = [
-        deliveryAddress.addressLine1,
-        deliveryAddress.addressLine2,
-        cityRegionFields.compactMap { $0 }.joined(),
-        deliveryAddress.countryCode.rawValue,
-        deliveryAddress.phoneNumber
-      ]
-      // Create address from all fields that are not nil and not the empty string.
-      return fields.compactMap { ($0 ?? "").isEmpty ? nil : $0 }.joined(separator: "\n")
-    }
-    let addressWithName = addressWithoutName.flatMap { address in
-      guard let name = backing?.deliveryAddress?.recipientName else {
-        return address
-      }
-      return name + "\n" + address
-    }
-
-    let alerts: [PPOProjectCardModel.Alert] = card.flags?
-      .compactMap { PPOProjectCardModel.Alert(flag: $0) } ?? []
-
-    let actionResult: (primary: Action, secondary: Action?, tier: TierType)?
-
-    switch card.tierType {
-    case PPOProjectCardModelConstants.paymentFailed:
-      actionResult = Self.actionsForPaymentFailed()
-    case PPOProjectCardModelConstants.confirmAddress:
-      actionResult = Self.actionsForConfirmAddress(address: addressWithoutName, addressId: addressId)
-    case PPOProjectCardModelConstants.completeSurvey:
-      actionResult = Self.actionsForSurvey()
-    case PPOProjectCardModelConstants.authenticationRequired:
-      actionResult = Self.actionsForAuthentication(clientSecret: backing?.clientSecret)
-    default:
-      return nil
-    }
-
-    guard let (primaryAction, secondaryAction, tierType) = actionResult else { return nil }
-
-    let projectAnalyticsFragment = backing?.project?.fragments.projectAnalyticsFragment
-
-    // For v1 of PPO we're just using the same url for surveys and the backing details page.
-    // This specifically links to the survey tab.
-    let backingDetailsUrl = backing?.backingDetailsPageRoute
-    let backingId = backing.flatMap { decompose(id: $0.id) }
-    let backingGraphId = backing?.id
-
-    if let image, let projectName, let projectId, let formattedPledge, let creatorName,
-       let projectAnalyticsFragment, let backingDetailsUrl, let backingId, let backingGraphId {
-      self.init(
-        isUnread: true,
-        alerts: alerts,
-        image: image,
-        projectName: projectName,
-        projectId: projectId,
-        pledge: formattedPledge,
-        creatorName: creatorName,
-        address: addressWithName,
-        actions: (primaryAction, secondaryAction),
-        tierType: tierType,
-        backingDetailsUrl: backingDetailsUrl,
-        backingId: backingId,
-        backingGraphId: backingGraphId,
-        projectAnalytics: projectAnalyticsFragment
-      )
-    } else {
-      return nil
-    }
-  }
-
-  private static func actionsForPaymentFailed() -> (Action, Action?, TierType) {
-    (.fixPayment, nil, .fixPayment)
-  }
-
-  private static func actionsForConfirmAddress(address: String?, addressId: String?)
-    -> (Action, Action?, TierType)? {
-    guard let address = address,
-          let addressId = addressId else { return nil }
-
-    return (
-      .confirmAddress(address: address, addressId: addressId),
-      .editAddress,
-      .confirmAddress
-    )
-  }
-
-  private static func actionsForSurvey() -> (Action, Action?, TierType) {
-    (.completeSurvey, nil, .openSurvey)
-  }
-
-  private static func actionsForAuthentication(clientSecret: String?)
-    -> (Action, Action?, TierType)? {
-    guard let clientSecret = clientSecret else { return nil }
-
-    return (
-      .authenticateCard(clientSecret: clientSecret),
-      nil,
-      .authenticateCard
-    )
-  }
-}
